[refactor] order 4계층 구조화
