--- order/src/main/java/com/sparta/rooibos/order/infrastructure/delivery/UpdateDeliveryResponse.java
@@ -1,19 +1,0 @@
-package com.sparta.rooibos.order.infrastructure.delivery;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-public record UpdateDeliveryResponse(
-    UUID deliveryId,
-    String status,
-    UUID departure,
-    UUID arrival,
-    String address,
-    UUID recipient,
-    String slackAccount,
-    UUID deliverId,
-    LocalDateTime updatedAt,
-    String updatedBy
-) {
-
-}
